+++ CHANGELOG.md
@@ -17,6 +17,9 @@ Add a link to the GitHub diff like
 
 [Unreleased]: https://github.com/mastersign/Mastersign.AutoForm/compare/main...dev
 
+### Fixed
+* Checkbox _Only current_ was always disabled
+
 ## [1.1.0] - 2022-02-24
 
 [1.1.0]: https://github.com/mastersign/Mastersign.AutoForm/compare/v1.0.0...v1.1.0