+++ CHANGELOG.md
@@ -17,6 +17,10 @@ Add a link to the GitHub diff like
 
 [Unreleased]: https://github.com/mastersign/Mastersign.AutoForm/compare/main...dev
 
+## [1.1.1] - 2022-02-24
+
+[1.1.1]: https://github.com/mastersign/Mastersign.AutoForm/compare/v1.1.0...v1.1.1
+
 ### Fixed
 * Checkbox _Only current_ was always disabled